requirements.txt
Django==2.2.6
djangorestframework==3.10.3
djangorestframework-bulk==0.2.1
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
django-cors-headers==3.1.1
django-filter==2.2.0
django-webpack-loader==0.6.0
docxtpl==0.9.0
easy-thumbnails==2.7
transliterate==1.10.2
django-simple-history==2.10.0
password_gen==0.2.17
django-richtextfield==1.6.1